ptch_docker/c_advanced/requirements.txt
beautifulsoup4==4.12.3
pillow==11.1.0
selenium==4.10.0
